Nest org_top_actions components under a namespace + add wrapper
Move the four registrations/show/org_top_actions_* panel components into a
nested org_top_actions/ directory (re-namespacing under OrgTopActions::) and
extract the action-panels controller div into an OrgTopActions::Wrapper
component rendered from org_admin.
