+++ app/components/registrations/show/org_top_actions/impound_update/component.en.yml
@@ -1,0 +1,8 @@
+---
+en:
+  components:
+    registrations:
+      show:
+        org_top_actions:
+          impound_update:
+            title: Update impound record
