+++ app/components/registrations/show/org_top_actions/parking_notification_form/component.en.yml
@@ -1,0 +1,12 @@
+---
+en:
+  components:
+    registrations:
+      show:
+        org_top_actions:
+          parking_notification_form:
+            address_or_intersection: Address or intersection
+            create_parking_notification: Create a new parking notification
+            impound_this_bike_type: Impound this %{bike_type}
+            parking_notification_notes_html: "%{org_name} internal notes <small>not
+              shown to user</small>"
