+++ app/components/registrations/show/org_top_actions/parking_notification_form/preview/default.html.erb
@@ -21,5 +21,5 @@
     Impound
   </button>
 
-  <%= render(Registrations::Show::OrgTopActionsParkingNotificationForm::Component.new(bike:, organization:)) %>
+  <%= render(Registrations::Show::OrgTopActions::ParkingNotificationForm::Component.new(bike:, organization:)) %>
 </div>
