+++ app/components/registrations/show/org_top_actions/wrapper/component.html.erb
@@ -1,0 +1,6 @@
+<div
+  class="tw:flex tw:flex-col tw:gap-3"
+  data-controller="registrations--show--action-panels"
+>
+  <%= content %>
+</div>
