+++ app/components/registrations/show/org_top_actions/wrapper/component.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Registrations
+  module Show
+    module OrgTopActions
+      module Wrapper
+        # Wraps the org-admin top action buttons and their accordion panels in the
+        # `registrations--show--action-panels` Stimulus controller
+        class Component < ApplicationComponent
+        end
+      end
+    end
+  end
+end
