--- app/components/registrations/show/org_top_actions_impound_update/component.en.yml
@@ -1,7 +1,0 @@
----
-en:
-  components:
-    registrations:
-      show:
-        org_top_actions_impound_update:
-          title: Update impound record
