--- app/components/registrations/show/org_top_actions_message_owner/component.en.yml
@@ -1,12 +1,0 @@
----
-en:
-  components:
-    registrations:
-      show:
-        org_top_actions_message_owner:
-          know_something_about_this_bike_type: Know something about this %{bike_type}?
-          link_url: Link url
-          send_message: Send message
-          where_did_you_see_this_bike: >-
-            Where did you see this %{bike_type}? If you're contacting about an ad
-            you found online, include a link in the field below
