--- app/components/registrations/show/org_top_actions_parking_notification_show/component.en.yml
@@ -1,10 +1,0 @@
----
-en:
-  components:
-    registrations:
-      show:
-        org_top_actions_parking_notification_show:
-          no_notifications: No parking notifications for this %{bike_type}
-          parking_notifications: Parking notifications
-          view_all: view all notifications for this %{bike_type}
-          view_notification: View notification
